Entities moved to separate files and foldered
